poster-engine-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
httpx==0.28.1
prometheus-client==0.21.1
orjson==3.10.12
pytest==8.3.5
PyJWT==2.10.1
boto3==1.38.19
